morph9.0 merge

+++ srchybrid/TransferWnd.cpp
@@ -1402,7 +1402,7 @@ void CTransferWnd::UpdateCatTabTitles(bool force)
 	if (!force && GetTabUnderMouse(&pt)!=-1)		// avoid cat tooltip jumping
 		return;
 
-	for (uint8 i=0;i<m_dlTab.GetItemCount();i++)
+	for (int i = 0;i < m_dlTab.GetItemCount();i++)
 		//MORPH START - Changed by SiRoB, Due to Khaos Category
 		//EditCatTabLabel(i,/*(i==0)? GetCatTitle( thePrefs.GetCategory(0)->filter ):*/thePrefs.GetCategory(i)->title);
 		EditCatTabLabel(i, thePrefs.GetCategory(i)->title);
@@ -1412,7 +1412,7 @@ void CTransferWnd::UpdateCatTabTitles(bool force)
 void CTransferWnd::EditCatTabLabel(int i)
 {
 	//MORPH	- Changed by SiRoB, Khaos Category
-	//	EditCatTabLabel(i,/*(i==0)? GetCatTitle( thePrefs.GetAllcatType() ):*/thePrefs.GetCategory(i)->title);
+	//EditCatTabLabel(i,/*(i==0)? GetCatTitle( thePrefs.GetAllcatType() ):*/thePrefs.GetCategory(i)->title);
 	EditCatTabLabel(i,thePrefs.GetCategory(i)->title);
 }
